feat: Refactor Dataspace folder
