+++ src/tractusx_sdk/dataspace/config/__init__.py
@@ -21,3 +21,5 @@
 #################################################################################
 
 ## Where the configuration of sdk is stored
+
+from .initial_config import auth_manager, edc_service, logger, args, app_configuration
